pyproject.toml
[project]
name = "akkurim-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "supertokens-python",
    "asyncpg",
    "pydantic",
    "pydantic-settings",
    "pydantic[email]",
    "pytest",
    "sse-starlette",
    "broadcaster",
    "orjson",
    "uvloop; sys_platform != 'win32'",
    "httptools",
    "fastapi-utils>=0.8.0",
    "alembic>=1.14.0",
    "psycopg[binary]>=3.2.3",
]
